+++ src/main/java/com/f1/quiket/domain/lecture/dto/LectureMaterialAiProcessRequest.java
@@ -1,5 +1,7 @@
 package com.f1.quiket.domain.lecture.dto;
 
+import com.f1.quiket.domain.material.dto.StudyMaterialFile;
+import com.f1.quiket.domain.material.dto.StudyMaterialUploadType;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,11 +12,11 @@ import lombok.Getter;
 @Getter
 @Builder
 public class LectureMaterialAiProcessRequest {
-    private final LectureUploadType uploadType;
+    private final StudyMaterialUploadType uploadType;
     private final PartSplitMethod partSplitMethod;
     private final String chapterName;
     private final String text;
-    private final List<LectureMaterialFile> files;
+    private final List<StudyMaterialFile> files;
     private final List<LecturePartSplitPlan> partSplitPlans;
 }
