--- src/main/java/com/f1/quiket/domain/lecture/dto/LecturePdfTextExtractionResult.java
@@ -1,15 +1,0 @@
-package com.f1.quiket.domain.lecture.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-/**
- * PDF 텍스트 추출 결과 DTO
- */
-@Getter
-@Builder
-public class LecturePdfTextExtractionResult {
-    private final boolean hasTextLayer;
-    private final String extractedText;
-}
-
