+++ src/main/java/com/f1/quiket/domain/material/dto/StudyMaterialFile.java
@@ -1,0 +1,16 @@
+package com.f1.quiket.domain.material.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+/**
+ * 강의 자료 파일 DTO
+ */
+@Getter
+@Builder
+public class StudyMaterialFile {
+    private final String fileName;
+    private final String contentType;
+    private final byte[] bytes;
+}
+
