+++ src/main/java/com/f1/quiket/domain/material/dto/StudyMaterialPdfTextExtractionResult.java
@@ -1,0 +1,15 @@
+package com.f1.quiket.domain.material.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+/**
+ * PDF 텍스트 추출 결과 DTO
+ */
+@Getter
+@Builder
+public class StudyMaterialPdfTextExtractionResult {
+    private final boolean hasTextLayer;
+    private final String extractedText;
+}
+
