--- src/main/java/com/f1/quiket/infra/ai/config/AiClientConfig.java
@@ -1,38 +1,0 @@
-package com.f1.quiket.infra.ai.config;
-
-import java.net.http.HttpClient;
-import java.time.Duration;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.JdkClientHttpRequestFactory;
-import org.springframework.web.client.RestClient;
-
-@Configuration
-@EnableConfigurationProperties({GeminiProperties.class, GroqProperties.class})
-public class AiClientConfig {
-
-    @Bean
-    @Qualifier("geminiRestClient")
-    public RestClient geminiRestClient(GeminiProperties properties) {
-        HttpClient httpClient = HttpClient.newBuilder()
-                .connectTimeout(Duration.ofSeconds(properties.getConnectTimeoutSeconds()))
-                .build();
-        JdkClientHttpRequestFactory requestFactory = new JdkClientHttpRequestFactory(httpClient);
-        requestFactory.setReadTimeout(Duration.ofSeconds(properties.getReadTimeoutSeconds()));
-        return RestClient.builder().requestFactory(requestFactory).build();
-    }
-
-    @Bean
-    @Qualifier("groqRestClient")
-    public RestClient groqRestClient(GroqProperties properties) {
-        HttpClient httpClient = HttpClient.newBuilder()
-                .connectTimeout(Duration.ofSeconds(properties.getConnectTimeoutSeconds()))
-                .build();
-        JdkClientHttpRequestFactory requestFactory = new JdkClientHttpRequestFactory(httpClient);
-        requestFactory.setReadTimeout(Duration.ofSeconds(properties.getReadTimeoutSeconds()));
-        return RestClient.builder().requestFactory(requestFactory).build();
-    }
-}
-
