+++ src/main/java/com/f1/quiket/infra/gemini/dto/GeminiCompletionResponse.java
@@ -1,0 +1,14 @@
+package com.f1.quiket.infra.gemini.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+/**
+ * 공통 텍스트 응답 DTO
+ */
+@Getter
+@Builder
+public class GeminiCompletionResponse {
+    private final String content;
+}
+
