+++ src/main/java/com/f1/quiket/infra/groq/dto/GroqCompletionResponse.java
@@ -1,0 +1,14 @@
+package com.f1.quiket.infra.groq.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+/**
+ * Groq 텍스트 응답 DTO
+ */
+@Getter
+@Builder
+public class GroqCompletionResponse {
+    private final String content;
+}
+
