+++ src/test/java/com/f1/quiket/domain/material/processor/StudyMaterialTextExtractorTest.java
@@ -1,4 +1,4 @@
-package com.f1.quiket.domain.lecture.service;
+package com.f1.quiket.domain.material.processor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -6,11 +6,10 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.f1.quiket.domain.lecture.dto.LectureMaterialAiProcessRequest;
-import com.f1.quiket.domain.lecture.dto.LectureMaterialAiProcessResult;
-import com.f1.quiket.domain.lecture.dto.PartSplitMethod;
 import com.f1.quiket.domain.material.dto.StudyMaterialFile;
 import com.f1.quiket.domain.material.dto.StudyMaterialPdfTextExtractionResult;
+import com.f1.quiket.domain.material.dto.StudyMaterialTextExtractionRequest;
+import com.f1.quiket.domain.material.dto.StudyMaterialTextExtractionResult;
 import com.f1.quiket.domain.material.dto.StudyMaterialUploadType;
 import com.f1.quiket.domain.material.port.StudyMaterialAiGateway;
 import com.f1.quiket.domain.material.port.StudyMaterialPdfTextExtractor;
@@ -19,52 +18,40 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-class LectureMaterialAiProcessorTest {
+class StudyMaterialTextExtractorTest {
 
     private StudyMaterialAiGateway studyMaterialAiGateway;
     private StudyMaterialPdfTextExtractor studyMaterialPdfTextExtractor;
-    private LectureMaterialAiProcessor lectureMaterialAiProcessor;
+    private StudyMaterialTextExtractor studyMaterialTextExtractor;
 
     @BeforeEach
     void setUp() {
         studyMaterialAiGateway = Mockito.mock(StudyMaterialAiGateway.class);
         studyMaterialPdfTextExtractor = Mockito.mock(StudyMaterialPdfTextExtractor.class);
-        lectureMaterialAiProcessor = new LectureMaterialAiProcessor(
+        studyMaterialTextExtractor = new StudyMaterialTextExtractor(
                 studyMaterialAiGateway,
                 studyMaterialPdfTextExtractor
         );
     }
 
     @Test
-    void process_text_uses_groq_and_parses_parts() {
-        when(studyMaterialAiGateway.generateFromText(any(), any()))
-                .thenReturn("""
-                        {
-                          "parts": [
-                            {"partNumber": 1, "name": "데이터 모델링 개념", "content": "데이터 모델링 기본 개념 정리"}
-                          ]
-                        }
-                        """);
-
-        LectureMaterialAiProcessResult result = lectureMaterialAiProcessor.process(
-                LectureMaterialAiProcessRequest.builder()
+    void extract_text_returns_input_without_ai_call() {
+        StudyMaterialTextExtractionResult result = studyMaterialTextExtractor.extract(
+                StudyMaterialTextExtractionRequest.builder()
                         .uploadType(StudyMaterialUploadType.TEXT)
-                        .partSplitMethod(PartSplitMethod.AUTO)
-                        .chapterName("1장 데이터 모델링")
-                        .text("데이터 모델링은 현실 세계의 데이터를 추상화한다.")
+                        .text("직접 입력한 파트 본문")
                         .build()
         );
 
-        assertThat(result.getProvider()).isEqualTo("groq");
-        assertThat(result.getParts()).hasSize(1);
-        assertThat(result.getParts().get(0).getName()).isEqualTo("데이터 모델링 개념");
-        verify(studyMaterialAiGateway).generateFromText(any(), any());
+        assertThat(result.getProvider()).isEqualTo("none");
+        assertThat(result.getExtractedText()).isEqualTo("직접 입력한 파트 본문");
+        verify(studyMaterialAiGateway, never()).generateFromText(any(), any());
         verify(studyMaterialAiGateway, never()).generateFromImages(any(), any(), any());
         verify(studyMaterialAiGateway, never()).generateFromPdf(any(), any(), any());
     }
 
     @Test
-    void process_pdf_with_text_layer_uses_groq() {
+    void extract_pdf_with_text_layer_uses_tika_text_without_ai_call() {
         StudyMaterialFile pdfFile = StudyMaterialFile.builder()
                 .fileName("lecture.pdf")
                 .contentType("application/pdf")
@@ -73,30 +60,23 @@ class LectureMaterialAiProcessorTest {
         when(studyMaterialPdfTextExtractor.extract(any()))
                 .thenReturn(StudyMaterialPdfTextExtractionResult.builder()
                         .hasTextLayer(true)
-                        .extractedText("텍스트 레이어 본문")
+                        .extractedText("PDF 텍스트 레이어 본문")
                         .build());
-        when(studyMaterialAiGateway.generateFromText(any(), any()))
-                .thenReturn("""
-                        {"parts":[{"partNumber":1,"name":"텍스트 기반 파트","content":"본문"}]}
-                        """);
 
-        LectureMaterialAiProcessResult result = lectureMaterialAiProcessor.process(
-                LectureMaterialAiProcessRequest.builder()
+        StudyMaterialTextExtractionResult result = studyMaterialTextExtractor.extract(
+                StudyMaterialTextExtractionRequest.builder()
                         .uploadType(StudyMaterialUploadType.PDF)
-                        .partSplitMethod(PartSplitMethod.AUTO)
-                        .chapterName("1장")
                         .files(List.of(pdfFile))
                         .build()
         );
 
-        assertThat(result.getProvider()).isEqualTo("groq");
-        assertThat(result.getExtractedText()).isEqualTo("텍스트 레이어 본문");
-        verify(studyMaterialAiGateway).generateFromText(any(), any());
+        assertThat(result.getProvider()).isEqualTo("tika");
+        assertThat(result.getExtractedText()).isEqualTo("PDF 텍스트 레이어 본문");
         verify(studyMaterialAiGateway, never()).generateFromPdf(any(), any(), any());
     }
 
     @Test
-    void process_pdf_without_text_layer_uses_gemini() {
+    void extract_pdf_without_text_layer_uses_gemini_ocr_only() {
         StudyMaterialFile pdfFile = StudyMaterialFile.builder()
                 .fileName("scan.pdf")
                 .contentType("application/pdf")
@@ -108,50 +88,39 @@ class LectureMaterialAiProcessorTest {
                         .extractedText("")
                         .build());
         when(studyMaterialAiGateway.generateFromPdf(any(), any(), any()))
-                .thenReturn("""
-                        {"parts":[{"partNumber":1,"partName":"OCR 파트","content":"OCR 본문"}]}
-                        """);
+                .thenReturn("스캔 PDF OCR 본문");
 
-        LectureMaterialAiProcessResult result = lectureMaterialAiProcessor.process(
-                LectureMaterialAiProcessRequest.builder()
+        StudyMaterialTextExtractionResult result = studyMaterialTextExtractor.extract(
+                StudyMaterialTextExtractionRequest.builder()
                         .uploadType(StudyMaterialUploadType.PDF)
-                        .partSplitMethod(PartSplitMethod.AUTO)
-                        .chapterName("1장")
                         .files(List.of(pdfFile))
                         .build()
         );
 
         assertThat(result.getProvider()).isEqualTo("gemini");
-        assertThat(result.getParts().get(0).getName()).isEqualTo("OCR 파트");
+        assertThat(result.getExtractedText()).isEqualTo("스캔 PDF OCR 본문");
         verify(studyMaterialAiGateway).generateFromPdf(any(), any(), any());
-        verify(studyMaterialAiGateway, never()).generateFromText(any(), any());
     }
 
     @Test
-    void process_image_parses_fenced_json_response() {
+    void extract_image_uses_gemini_ocr_only() {
         StudyMaterialFile imageFile = StudyMaterialFile.builder()
                 .fileName("slide.png")
                 .contentType("image/png")
                 .bytes("image".getBytes())
                 .build();
         when(studyMaterialAiGateway.generateFromImages(any(), any(), any()))
-                .thenReturn("""
-                        ```json
-                        {"parts":[{"partNumber":1,"name":"이미지 파트","content":"이미지 OCR 본문"}]}
-                        ```
-                        """);
+                .thenReturn("이미지 OCR 본문");
 
-        LectureMaterialAiProcessResult result = lectureMaterialAiProcessor.process(
-                LectureMaterialAiProcessRequest.builder()
+        StudyMaterialTextExtractionResult result = studyMaterialTextExtractor.extract(
+                StudyMaterialTextExtractionRequest.builder()
                         .uploadType(StudyMaterialUploadType.IMAGE)
-                        .partSplitMethod(PartSplitMethod.AUTO)
-                        .chapterName("1장")
                         .files(List.of(imageFile))
                         .build()
         );
 
         assertThat(result.getProvider()).isEqualTo("gemini");
-        assertThat(result.getParts().get(0).getContent()).isEqualTo("이미지 OCR 본문");
+        assertThat(result.getExtractedText()).isEqualTo("이미지 OCR 본문");
         verify(studyMaterialAiGateway).generateFromImages(any(), any(), any());
     }
 }
